22.2.8

+++ bgimage/sync.js
@@ -10,33 +10,39 @@ Sync.BGImage = {
 		delete Sync.BGImage.tid[Ctrl.Id];
 		const hwnd = Ctrl.hwndList;
 		if (hwnd) {
-			const r = Sync.BGImage.db[hwnd];
-			if ("object" === typeof r) {
-				Sync.BGImage.ShowImage(hwnd);
-				return;
-			}
-			if (r === 1) {
-				return;
-			}
-			Sync.BGImage.db[hwnd] = 1;
 			const path = Ctrl.FolderItem.Path;
+			let db = Sync.BGImage.db[path];
+			if (db) {
+				db[hwnd] = 1;
+				if (db.image) {
+					Sync.BGImage.ShowImage(db);
+				}
+				return;
+			}
+			db = {};
+			db[hwnd] = 1;
+			Sync.BGImage.db[path] = db;
 			const list = Sync.BGImage.List;
 			for (let i = list.length; i--;) {
 				if (PathMatchEx(path, list[i][0])) {
 					const image = list[i][1];
 					if ("object" === typeof image) {
-						Sync.BGImage.ShowImage(hwnd, image);
-					} else if ("string" === typeof image) {
+						db.image = image;
+						Sync.BGImage.ShowImage(db);
+						break;
+					}
+					if ("string" === typeof image) {
 						Threads.GetImage({
-							hwnd: hwnd,
 							path: image,
+							db: db,
 							listx: list[i],
 
 							onload: function (o) {
 								if ("object" !== typeof o.listx[1]) {
 									o.listx[1] = o.out;
 								}
-								Sync.BGImage.ShowImage(hwnd, o.listx[1]);
+								o.db.image = o.listx[1];
+								Sync.BGImage.ShowImage(o.db);
 							},
 
 							onerror: function (o) {
@@ -52,13 +58,17 @@ Sync.BGImage = {
 		}
 	},
 
-	ShowImage: function (hwnd, image) {
-		Sync.BGImage.db[hwnd] = image;
-		const lvbk = api.Memory("LVBKIMAGE");
-		lvbk.ulFlags = LVBKIF_TYPE_WATERMARK | LVBKIF_FLAG_ALPHABLEND;
-		lvbk.hbm = image.GetHBITMAP(-2);
-		if (!api.SendMessage(hwnd, LVM_SETBKIMAGE, 0, lvbk)) {
-			api.DeleteObject(lvbk.hbm);
+	ShowImage: function (db) {
+		for (hwnd in db) {
+			if (hwnd !== "image") {
+				delete db[hwnd];
+				const lvbk = api.Memory("LVBKIMAGE");
+				lvbk.ulFlags = LVBKIF_TYPE_WATERMARK | LVBKIF_FLAG_ALPHABLEND;
+				lvbk.hbm = db.image.GetHBITMAP(-2);
+				if (!api.SendMessage(hwnd, LVM_SETBKIMAGE, 0, lvbk)) {
+					api.DeleteObject(lvbk.hbm);
+				}
+			}
 		}
 	},
